third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/fullscreen-elements-do-not-affect-modality-expected.txt

Tests that the active modal dialog is as expected when fullscreen elements are added to the top layer. This test requires testRunner.

the active modal dialog should be: top-dialog
EXPECTED (clickedOn == '[object HTMLDialogElement]') OK
EXPECTED (clickedOn == '[object HTMLHtmlElement]') OK
the active modal dialog should be: bottom-dialog
EXPECTED (clickedOn == '[object HTMLDialogElement]') OK
EXPECTED (clickedOn == '[object HTMLHtmlElement]') OK
there should be no modal dialog
EXPECTED (clickedOn == '[object HTMLDivElement]') OK
END OF TEST
